move distribution to "jfrog ds"
