+++ artifactory/cli.go
@@ -8,7 +8,6 @@ import (
 	"github.com/jfrog/jfrog-cli/docs/artifactory/yarnconfig"
 	"io/ioutil"
 	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
 
@@ -45,7 +44,6 @@ import (
 	"github.com/codegangsta/cli"
 	"github.com/jfrog/jfrog-cli-core/artifactory/commands/buildinfo"
 	"github.com/jfrog/jfrog-cli-core/artifactory/commands/curl"
-	"github.com/jfrog/jfrog-cli-core/artifactory/commands/distribution"
 	"github.com/jfrog/jfrog-cli-core/artifactory/commands/generic"
 	"github.com/jfrog/jfrog-cli-core/artifactory/commands/golang"
 	"github.com/jfrog/jfrog-cli-core/artifactory/commands/gradle"
@@ -100,11 +98,6 @@ import (
 	"github.com/jfrog/jfrog-cli/docs/artifactory/ping"
 	"github.com/jfrog/jfrog-cli/docs/artifactory/pipconfig"
 	"github.com/jfrog/jfrog-cli/docs/artifactory/pipinstall"
-	"github.com/jfrog/jfrog-cli/docs/artifactory/releasebundlecreate"
-	"github.com/jfrog/jfrog-cli/docs/artifactory/releasebundledelete"
-	"github.com/jfrog/jfrog-cli/docs/artifactory/releasebundledistribute"
-	"github.com/jfrog/jfrog-cli/docs/artifactory/releasebundlesign"
-	"github.com/jfrog/jfrog-cli/docs/artifactory/releasebundleupdate"
 	"github.com/jfrog/jfrog-cli/docs/artifactory/replicationcreate"
 	"github.com/jfrog/jfrog-cli/docs/artifactory/replicationdelete"
 	"github.com/jfrog/jfrog-cli/docs/artifactory/replicationtemplate"
@@ -120,8 +113,6 @@ import (
 	"github.com/jfrog/jfrog-cli/utils/cliutils"
 	buildinfocmd "github.com/jfrog/jfrog-client-go/artifactory/buildinfo"
 	"github.com/jfrog/jfrog-client-go/artifactory/services"
-	distributionServices "github.com/jfrog/jfrog-client-go/distribution/services"
-	distributionServicesUtils "github.com/jfrog/jfrog-client-go/distribution/services/utils"
 	clientutils "github.com/jfrog/jfrog-client-go/utils"
 	"github.com/jfrog/jfrog-client-go/utils/errorutils"
 	"github.com/jfrog/jfrog-client-go/utils/log"
@@ -757,71 +748,6 @@ func GetCommands() []cli.Command {
 			},
 		},
 		{
-			Name:         "release-bundle-create",
-			Flags:        cliutils.GetCommandFlags(cliutils.ReleaseBundleCreate),
-			Aliases:      []string{"rbc"},
-			Description:  releasebundlecreate.Description,
-			HelpName:     corecommon.CreateUsage("rt rbc", releasebundlecreate.Description, releasebundlecreate.Usage),
-			UsageText:    releasebundlecreate.Arguments,
-			ArgsUsage:    common.CreateEnvVars(),
-			BashComplete: corecommon.CreateBashCompletionFunc(),
-			Action: func(c *cli.Context) error {
-				return releaseBundleCreateCmd(c)
-			},
-		},
-		{
-			Name:         "release-bundle-update",
-			Flags:        cliutils.GetCommandFlags(cliutils.ReleaseBundleUpdate),
-			Aliases:      []string{"rbu"},
-			Description:  releasebundleupdate.Description,
-			HelpName:     corecommon.CreateUsage("rt rbu", releasebundleupdate.Description, releasebundleupdate.Usage),
-			UsageText:    releasebundleupdate.Arguments,
-			ArgsUsage:    common.CreateEnvVars(),
-			BashComplete: corecommon.CreateBashCompletionFunc(),
-			Action: func(c *cli.Context) error {
-				return releaseBundleUpdateCmd(c)
-			},
-		},
-		{
-			Name:         "release-bundle-sign",
-			Flags:        cliutils.GetCommandFlags(cliutils.ReleaseBundleSign),
-			Aliases:      []string{"rbs"},
-			Description:  releasebundlesign.Description,
-			HelpName:     corecommon.CreateUsage("rt rbs", releasebundlesign.Description, releasebundlesign.Usage),
-			UsageText:    releasebundlesign.Arguments,
-			ArgsUsage:    common.CreateEnvVars(),
-			BashComplete: corecommon.CreateBashCompletionFunc(),
-			Action: func(c *cli.Context) error {
-				return releaseBundleSignCmd(c)
-			},
-		},
-		{
-			Name:         "release-bundle-distribute",
-			Flags:        cliutils.GetCommandFlags(cliutils.ReleaseBundleDistribute),
-			Aliases:      []string{"rbd"},
-			Description:  releasebundledistribute.Description,
-			HelpName:     corecommon.CreateUsage("rt rbd", releasebundledistribute.Description, releasebundledistribute.Usage),
-			UsageText:    releasebundledistribute.Arguments,
-			ArgsUsage:    common.CreateEnvVars(),
-			BashComplete: corecommon.CreateBashCompletionFunc(),
-			Action: func(c *cli.Context) error {
-				return releaseBundleDistributeCmd(c)
-			},
-		},
-		{
-			Name:         "release-bundle-delete",
-			Flags:        cliutils.GetCommandFlags(cliutils.ReleaseBundleDelete),
-			Aliases:      []string{"rbdel"},
-			Description:  releasebundledelete.Description,
-			HelpName:     corecommon.CreateUsage("rt rbdel", releasebundledelete.Description, releasebundledelete.Usage),
-			UsageText:    releasebundledelete.Arguments,
-			ArgsUsage:    common.CreateEnvVars(),
-			BashComplete: corecommon.CreateBashCompletionFunc(),
-			Action: func(c *cli.Context) error {
-				return releaseBundleDeleteCmd(c)
-			},
-		},
-		{
 			Name:         "repo-template",
 			Aliases:      []string{"rpt"},
 			Description:  repotemplate.Description,
@@ -1053,7 +979,7 @@ func GetCommands() []cli.Command {
 	})
 }
 
-func createArtifactoryDetailsByFlags(c *cli.Context, distribution bool) (*coreConfig.ServerDetails, error) {
+func CreateArtifactoryDetailsByFlags(c *cli.Context, distribution bool) (*coreConfig.ServerDetails, error) {
 	artDetails, err := createArtifactoryDetailsWithConfigOffer(c, distribution)
 	if err != nil {
 		return nil, err
@@ -1367,7 +1293,7 @@ func dockerPromoteCmd(c *cli.Context) error {
 	if c.NArg() != 3 {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
-	artDetails, err := createArtifactoryDetailsByFlags(c, false)
+	artDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1386,7 +1312,7 @@ func containerPushCmd(c *cli.Context, containerManagerType containerutils.Contai
 	if c.NArg() != 2 {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
-	artDetails, err := createArtifactoryDetailsByFlags(c, false)
+	artDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1420,7 +1346,7 @@ func containerPullCmd(c *cli.Context, containerManagerType containerutils.Contai
 	if c.NArg() != 2 {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
-	artDetails, err := createArtifactoryDetailsByFlags(c, false)
+	artDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1441,7 +1367,7 @@ func BuildDockerCreateCmd(c *cli.Context) error {
 	if c.NArg() != 1 {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
-	artDetails, err := createArtifactoryDetailsByFlags(c, false)
+	artDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1520,7 +1446,7 @@ func nugetLegacyCmd(c *cli.Context) error {
 	if err != nil {
 		return err
 	}
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1623,7 +1549,7 @@ func npmLegacyInstallCmd(c *cli.Context) error {
 		return err
 	}
 	npmCmd := npm.NewNpmLegacyInstallCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1675,7 +1601,7 @@ func npmLegacyCiCmd(c *cli.Context) error {
 		return err
 	}
 	npmCmd := npm.NewNpmLegacyCiCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1730,7 +1656,7 @@ func npmLegacyPublishCmd(c *cli.Context) error {
 		return err
 	}
 	npmPublicCmd := npm.NewNpmPublishCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1876,7 +1802,7 @@ func goLegacyCmd(c *cli.Context) error {
 		err = cliutils.PrintSummaryReport(0, 1, err)
 	}
 	targetRepo := c.Args().Get(1)
-	details, err := createArtifactoryDetailsByFlags(c, false)
+	details, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -1931,7 +1857,7 @@ func goLegacyPublishCmd(c *cli.Context) error {
 	}
 	targetRepo := c.Args().Get(0)
 	version := c.Args().Get(1)
-	details, err := createArtifactoryDetailsByFlags(c, false)
+	details, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2018,7 +1944,7 @@ func pingCmd(c *cli.Context) error {
 	if c.NArg() > 0 {
 		return cliutils.PrintHelpAndReturnError("No arguments should be sent.", c)
 	}
-	artDetails, err := createArtifactoryDetailsByFlags(c, false)
+	artDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2045,7 +1971,7 @@ func prepareDownloadCommand(c *cli.Context) (*spec.SpecFiles, error) {
 	var downloadSpec *spec.SpecFiles
 	var err error
 	if c.IsSet("spec") {
-		downloadSpec, err = getSpec(c, true)
+		downloadSpec, err = GetSpec(c, true)
 	} else {
 		downloadSpec, err = createDefaultDownloadSpec(c)
 	}
@@ -2069,7 +1995,7 @@ func downloadCmd(c *cli.Context) error {
 	if err != nil {
 		return err
 	}
-	serverDetails, err := createArtifactoryDetailsByFlags(c, false)
+	serverDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2124,7 +2050,7 @@ func uploadCmd(c *cli.Context) error {
 		return err
 	}
 	uploadCmd := generic.NewUploadCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2171,7 +2097,7 @@ func prepareCopyMoveCommand(c *cli.Context) (*spec.SpecFiles, error) {
 	var copyMoveSpec *spec.SpecFiles
 	var err error
 	if c.IsSet("spec") {
-		copyMoveSpec, err = getSpec(c, false)
+		copyMoveSpec, err = GetSpec(c, false)
 	} else {
 		copyMoveSpec, err = createDefaultCopyMoveSpec(c)
 	}
@@ -2191,7 +2117,7 @@ func moveCmd(c *cli.Context) error {
 		return err
 	}
 	moveCmd := generic.NewMoveCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2214,7 +2140,7 @@ func copyCmd(c *cli.Context) error {
 	}
 
 	copyCommand := generic.NewCopyCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2241,7 +2167,7 @@ func prepareDeleteCommand(c *cli.Context) (*spec.SpecFiles, error) {
 	var deleteSpec *spec.SpecFiles
 	var err error
 	if c.IsSet("spec") {
-		deleteSpec, err = getSpec(c, false)
+		deleteSpec, err = GetSpec(c, false)
 	} else {
 		deleteSpec, err = createDefaultDeleteSpec(c)
 	}
@@ -2262,7 +2188,7 @@ func deleteCmd(c *cli.Context) error {
 	}
 
 	deleteCommand := generic.NewDeleteCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2290,7 +2216,7 @@ func prepareSearchCommand(c *cli.Context) (*spec.SpecFiles, error) {
 	var searchSpec *spec.SpecFiles
 	var err error
 	if c.IsSet("spec") {
-		searchSpec, err = getSpec(c, false)
+		searchSpec, err = GetSpec(c, false)
 	} else {
 		searchSpec, err = createDefaultSearchSpec(c)
 	}
@@ -2309,7 +2235,7 @@ func searchCmd(c *cli.Context) error {
 	if err != nil {
 		return err
 	}
-	artDetails, err := createArtifactoryDetailsByFlags(c, false)
+	artDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2349,7 +2275,7 @@ func preparePropsCmd(c *cli.Context) (*generic.PropsCommand, error) {
 	var props string
 	if c.IsSet("spec") {
 		props = c.Args()[0]
-		propsSpec, err = getSpec(c, false)
+		propsSpec, err = GetSpec(c, false)
 	} else {
 		propsSpec, err = createDefaultPropertiesSpec(c)
 		if c.NArg() == 1 {
@@ -2368,7 +2294,7 @@ func preparePropsCmd(c *cli.Context) (*generic.PropsCommand, error) {
 	}
 
 	command := generic.NewPropsCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return nil, err
 	}
@@ -2419,7 +2345,7 @@ func buildPublishCmd(c *cli.Context) error {
 		return err
 	}
 	buildInfoConfiguration := createBuildInfoConfiguration(c)
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2443,7 +2369,7 @@ func buildAppendCmd(c *cli.Context) error {
 		return err
 	}
 	buildNameToAppend, buildNumberToAppend := c.Args().Get(2), c.Args().Get(3)
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2480,7 +2406,7 @@ func buildAddDependenciesCmd(c *cli.Context) error {
 		dependenciesSpec = createDefaultBuildAddDependenciesSpec(c)
 	}
 	if c.Bool("from-rt") {
-		rtDetails, err = createArtifactoryDetailsByFlags(c, false)
+		rtDetails, err = CreateArtifactoryDetailsByFlags(c, false)
 		if err != nil {
 			return err
 		}
@@ -2537,7 +2463,7 @@ func buildScanCmd(c *cli.Context) error {
 	if err := validateBuildConfiguration(c, buildConfiguration); err != nil {
 		return err
 	}
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2580,7 +2506,7 @@ func buildPromoteCmd(c *cli.Context) error {
 		return err
 	}
 	configuration := createBuildPromoteConfiguration(c)
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2597,7 +2523,7 @@ func buildDistributeCmd(c *cli.Context) error {
 		return err
 	}
 	configuration := createBuildDistributionConfiguration(c)
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2615,7 +2541,7 @@ func buildDiscardCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Build name is expected as a command argument or environment variable.", c)
 	}
 	buildDiscardCmd := buildinfo.NewBuildDiscardCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2624,186 +2550,13 @@ func buildDiscardCmd(c *cli.Context) error {
 	return commands.Exec(buildDiscardCmd)
 }
 
-func releaseBundleCreateCmd(c *cli.Context) error {
-	if !(c.NArg() == 2 && c.IsSet("spec") || (c.NArg() == 3 && !c.IsSet("spec"))) {
-		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
-	}
-	if c.IsSet("detailed-summary") && !c.IsSet("sign") {
-		return cliutils.PrintHelpAndReturnError("The --detailed-summary option can't be used without --sign", c)
-	}
-	var releaseBundleCreateSpec *spec.SpecFiles
-	var err error
-	if c.IsSet("spec") {
-		releaseBundleCreateSpec, err = getSpec(c, true)
-	} else {
-		releaseBundleCreateSpec = createDefaultReleaseBundleSpec(c)
-	}
-	if err != nil {
-		return err
-	}
-	err = spec.ValidateSpec(releaseBundleCreateSpec.Files, false, true, false)
-	if err != nil {
-		return err
-	}
-
-	params, err := createReleaseBundleCreateUpdateParams(c, c.Args().Get(0), c.Args().Get(1))
-	if err != nil {
-		return err
-	}
-	releaseBundleCreateCmd := distribution.NewReleaseBundleCreateCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, true)
-	if err != nil {
-		return err
-	}
-	releaseBundleCreateCmd.SetServerDetails(rtDetails).SetReleaseBundleCreateParams(params).SetSpec(releaseBundleCreateSpec).SetDryRun(c.Bool("dry-run")).SetDetailedSummary(c.Bool("detailed-summary"))
-
-	commands.Exec(releaseBundleCreateCmd)
-	if releaseBundleCreateCmd.IsDetailedSummary() {
-		if summary := releaseBundleCreateCmd.GetSummary(); summary != nil {
-			return cliutils.PrintBuildInfoSummaryReport(summary.IsSucceeded(), summary.GetSha256(), err)
-		}
-	}
-	return err
-}
-
-func releaseBundleUpdateCmd(c *cli.Context) error {
-	if !(c.NArg() == 2 && c.IsSet("spec") || (c.NArg() == 3 && !c.IsSet("spec"))) {
-		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
-	}
-	if c.IsSet("detailed-summary") && !c.IsSet("sign") {
-		return cliutils.PrintHelpAndReturnError("The --detailed-summary option can't be used without --sign", c)
-	}
-	var releaseBundleUpdateSpec *spec.SpecFiles
-	var err error
-	if c.IsSet("spec") {
-		releaseBundleUpdateSpec, err = getSpec(c, true)
-	} else {
-		releaseBundleUpdateSpec = createDefaultReleaseBundleSpec(c)
-	}
-	if err != nil {
-		return err
-	}
-	err = spec.ValidateSpec(releaseBundleUpdateSpec.Files, false, true, false)
-	if err != nil {
-		return err
-	}
-
-	params, err := createReleaseBundleCreateUpdateParams(c, c.Args().Get(0), c.Args().Get(1))
-	if err != nil {
-		return err
-	}
-	releaseBundleUpdateCmd := distribution.NewReleaseBundleUpdateCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, true)
-	if err != nil {
-		return err
-	}
-	releaseBundleUpdateCmd.SetServerDetails(rtDetails).SetReleaseBundleUpdateParams(params).SetSpec(releaseBundleUpdateSpec).SetDryRun(c.Bool("dry-run")).SetDetailedSummary(c.Bool("detailed-summary"))
-
-	err = commands.Exec(releaseBundleUpdateCmd)
-	if releaseBundleUpdateCmd.IsDetailedSummary() {
-		if summary := releaseBundleUpdateCmd.GetSummary(); summary != nil {
-			return cliutils.PrintBuildInfoSummaryReport(summary.IsSucceeded(), summary.GetSha256(), err)
-		}
-	}
-	return err
-}
-
-func releaseBundleSignCmd(c *cli.Context) error {
-	if c.NArg() != 2 {
-		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
-	}
-
-	params := distributionServices.NewSignBundleParams(c.Args().Get(0), c.Args().Get(1))
-	params.StoringRepository = c.String("repo")
-	params.GpgPassphrase = c.String("passphrase")
-	releaseBundleSignCmd := distribution.NewReleaseBundleSignCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, true)
-	if err != nil {
-		return err
-	}
-	releaseBundleSignCmd.SetServerDetails(rtDetails).SetReleaseBundleSignParams(params).SetDetailedSummary(c.Bool("detailed-summary"))
-	err = commands.Exec(releaseBundleSignCmd)
-	if releaseBundleSignCmd.IsDetailedSummary() {
-		if summary := releaseBundleSignCmd.GetSummary(); summary != nil {
-			return cliutils.PrintBuildInfoSummaryReport(summary.IsSucceeded(), summary.GetSha256(), err)
-		}
-	}
-	return err
-}
-
-func releaseBundleDistributeCmd(c *cli.Context) error {
-	if c.NArg() != 2 {
-		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
-	}
-	if c.IsSet("max-wait-minutes") && !c.IsSet("sync") {
-		return cliutils.PrintHelpAndReturnError("The --max-wait-minutes option can't be used without --sync", c)
-	}
-	var distributionRules *spec.DistributionRules
-	if c.IsSet("dist-rules") {
-		if c.IsSet("site") || c.IsSet("city") || c.IsSet("country-code") {
-			return cliutils.PrintHelpAndReturnError("The --dist-rules option can't be used with --site, --city or --country-code", c)
-		}
-		var err error
-		distributionRules, err = spec.CreateDistributionRulesFromFile(c.String("dist-rules"))
-		if err != nil {
-			return err
-		}
-	} else {
-		distributionRules = createDefaultDistributionRules(c)
-	}
-
-	params := distributionServices.NewDistributeReleaseBundleParams(c.Args().Get(0), c.Args().Get(1))
-	releaseBundleDistributeCmd := distribution.NewReleaseBundleDistributeCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, true)
-	if err != nil {
-		return err
-	}
-	maxWaitMinutes, err := cliutils.GetIntFlagValue(c, "max-wait-minutes", 0)
-	if err != nil {
-		return err
-	}
-	releaseBundleDistributeCmd.SetServerDetails(rtDetails).SetDistributeBundleParams(params).SetDistributionRules(distributionRules).SetDryRun(c.Bool("dry-run")).SetSync(c.Bool("sync")).SetMaxWaitMinutes(maxWaitMinutes)
-
-	return commands.Exec(releaseBundleDistributeCmd)
-}
-
-func releaseBundleDeleteCmd(c *cli.Context) error {
-	if c.NArg() != 2 {
-		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
-	}
-	var distributionRules *spec.DistributionRules
-	if c.IsSet("dist-rules") {
-		if c.IsSet("site") || c.IsSet("city") || c.IsSet("country-code") {
-			return cliutils.PrintHelpAndReturnError("flag --dist-rules can't be used with --site, --city or --country-code", c)
-		}
-		var err error
-		distributionRules, err = spec.CreateDistributionRulesFromFile(c.String("dist-rules"))
-		if err != nil {
-			return err
-		}
-	} else {
-		distributionRules = createDefaultDistributionRules(c)
-	}
-
-	params := distributionServices.NewDeleteReleaseBundleParams(c.Args().Get(0), c.Args().Get(1))
-	params.DeleteFromDistribution = c.BoolT("delete-from-dist")
-	distributeBundleCmd := distribution.NewReleaseBundleDeleteParams()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, true)
-	if err != nil {
-		return err
-	}
-	distributeBundleCmd.SetQuiet(cliutils.GetQuietValue(c)).SetServerDetails(rtDetails).SetDistributeBundleParams(params).SetDistributionRules(distributionRules).SetDryRun(c.Bool("dry-run"))
-
-	return commands.Exec(distributeBundleCmd)
-}
-
 func gitLfsCleanCmd(c *cli.Context) error {
 	if c.NArg() > 1 {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 	configuration := createGitLfsCleanConfiguration(c)
 	gitLfsCmd := generic.NewGitLfsCommand()
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2879,7 +2632,7 @@ func repoCreateCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2895,7 +2648,7 @@ func repoUpdateCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2911,7 +2664,7 @@ func repoDeleteCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2934,7 +2687,7 @@ func replicationCreateCmd(c *cli.Context) error {
 	if c.NArg() != 1 {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2947,7 +2700,7 @@ func replicationDeleteCmd(c *cli.Context) error {
 	if c.NArg() != 1 {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2972,7 +2725,7 @@ func permissionTargetCreateCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -2988,7 +2741,7 @@ func permissionTargetUpdateCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3004,7 +2757,7 @@ func permissionTargetDeleteCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3019,7 +2772,7 @@ func userCreateCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3048,7 +2801,7 @@ func usersCreateCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3079,7 +2832,7 @@ func usersDeleteCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3140,7 +2893,7 @@ func groupCreateCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3156,7 +2909,7 @@ func groupAddUsersCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3172,7 +2925,7 @@ func groupDeleteCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	rtDetails, err := createArtifactoryDetailsByFlags(c, false)
+	rtDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3193,7 +2946,7 @@ func accessTokenCreateCmd(c *cli.Context) error {
 		return cliutils.PrintHelpAndReturnError("Wrong number of arguments.", c)
 	}
 
-	serverDetails, err := createArtifactoryDetailsByFlags(c, false)
+	serverDetails, err := CreateArtifactoryDetailsByFlags(c, false)
 	if err != nil {
 		return err
 	}
@@ -3334,7 +3087,7 @@ func createDefaultCopyMoveSpec(c *cli.Context) (*spec.SpecFiles, error) {
 		BuildSpec(), nil
 }
 
-func getSpec(c *cli.Context, isDownload bool) (specFiles *spec.SpecFiles, err error) {
+func GetSpec(c *cli.Context, isDownload bool) (specFiles *spec.SpecFiles, err error) {
 	specFiles, err = spec.CreateSpecFromFile(c.String("spec"), coreutils.SpecVarsStringToMap(c.String("spec-vars")))
 	if err != nil {
 		return nil, err
@@ -3487,26 +3240,6 @@ func createBuildDistributionConfiguration(c *cli.Context) services.BuildDistribu
 	return distributeParamsImpl
 }
 
-func createReleaseBundleCreateUpdateParams(c *cli.Context, bundleName, bundleVersion string) (distributionServicesUtils.ReleaseBundleParams, error) {
-	releaseBundleParams := distributionServicesUtils.NewReleaseBundleParams(bundleName, bundleVersion)
-	releaseBundleParams.SignImmediately = c.Bool("sign")
-	releaseBundleParams.StoringRepository = c.String("repo")
-	releaseBundleParams.GpgPassphrase = c.String("passphrase")
-	releaseBundleParams.Description = c.String("desc")
-	if c.IsSet("release-notes-path") {
-		bytes, err := ioutil.ReadFile(c.String("release-notes-path"))
-		if err != nil {
-			return releaseBundleParams, errorutils.CheckError(err)
-		}
-		releaseBundleParams.ReleaseNotes = string(bytes)
-		releaseBundleParams.ReleaseNotesSyntax, err = populateReleaseNotesSyntax(c)
-		if err != nil {
-			return releaseBundleParams, err
-		}
-	}
-	return releaseBundleParams, nil
-}
-
 func createGitLfsCleanConfiguration(c *cli.Context) (gitLfsCleanConfiguration *generic.GitLfsCleanConfiguration) {
 	gitLfsCleanConfiguration = new(generic.GitLfsCleanConfiguration)
 
@@ -3618,30 +3351,6 @@ func createDefaultBuildAddDependenciesSpec(c *cli.Context) *spec.SpecFiles {
 		Regexp(c.Bool("regexp")).
 		Ant(c.Bool("ant")).
 		BuildSpec()
-}
-
-func createDefaultReleaseBundleSpec(c *cli.Context) *spec.SpecFiles {
-	return spec.NewBuilder().
-		Pattern(c.Args().Get(2)).
-		Target(c.String("target")).
-		Props(c.String("props")).
-		Build(c.String("build")).
-		Bundle(c.String("bundle")).
-		Exclusions(cliutils.GetStringsArrFlagValue(c, "exclusions")).
-		Regexp(c.Bool("regexp")).
-		TargetProps(c.String("target-props")).
-		Ant(c.Bool("ant")).
-		BuildSpec()
-}
-
-func createDefaultDistributionRules(c *cli.Context) *spec.DistributionRules {
-	return &spec.DistributionRules{
-		DistributionRules: []spec.DistributionRule{{
-			SiteName:     c.String("site"),
-			CityName:     c.String("city"),
-			CountryCodes: cliutils.GetStringsArrFlagValue(c, "country-codes"),
-		}},
-	}
 }
 
 func getFileSystemSpec(c *cli.Context) (fsSpec *spec.SpecFiles, err error) {
@@ -3840,27 +3549,4 @@ func extractThreadsFlag(args []string) (cleanArgs []string, threadsCount int, er
 	}
 
 	return
-}
-
-func populateReleaseNotesSyntax(c *cli.Context) (distributionServicesUtils.ReleaseNotesSyntax, error) {
-	// If release notes syntax is set, use it
-	releaseNotexSyntax := c.String("release-notes-syntax")
-	if releaseNotexSyntax != "" {
-		switch releaseNotexSyntax {
-		case "markdown":
-			return distributionServicesUtils.Markdown, nil
-		case "asciidoc":
-			return distributionServicesUtils.Asciidoc, nil
-		case "plain_text":
-			return distributionServicesUtils.PlainText, nil
-		default:
-			return distributionServicesUtils.PlainText, errorutils.CheckError(errors.New("--release-notes-syntax must be one of: markdown, asciidoc or plain_text."))
-		}
-	}
-	// If the file extension is ".md" or ".markdown", use the markdonwn syntax
-	extension := strings.ToLower(filepath.Ext(c.String("release-notes-path")))
-	if extension == ".md" || extension == ".markdown" {
-		return distributionServicesUtils.Markdown, nil
-	}
-	return distributionServicesUtils.PlainText, nil
 }
