+++ docs/artifactory/releasebundlecreate/help.go
@@ -2,8 +2,8 @@ package releasebundlecreate
 
 const Description = "Create a release bundle."
 
-var Usage = []string{"jfrog rt rbc [command options] <release bundle name> <release bundle version> <pattern>",
-	"jfrog rt rbc --spec=<File Spec path> [command options] <release bundle name> <release bundle version>"}
+var Usage = []string{"jfrog ds rbc [command options] <release bundle name> <release bundle version> <pattern>",
+	"jfrog ds rbc --spec=<File Spec path> [command options] <release bundle name> <release bundle version>"}
 
 const Arguments string = `	release bundle name
 		The name of the release bundle.
