+++ docs/artifactory/releasebundledelete/help.go
@@ -2,7 +2,7 @@ package releasebundledelete
 
 const Description = "Delete a release bundle."
 
-var Usage = []string{"jfrog rt rbdel [command options] <release bundle name> <release bundle version>"}
+var Usage = []string{"jfrog ds rbdel [command options] <release bundle name> <release bundle version>"}
 
 const Arguments string = `	release bundle name
 		Release bundle name.
