+++ docs/artifactory/releasebundledistribute/help.go
@@ -2,7 +2,7 @@ package releasebundledistribute
 
 const Description = "Distribute a release bundle."
 
-var Usage = []string{"jfrog rt rbd [command options] <release bundle name> <release bundle version>"}
+var Usage = []string{"jfrog ds rbd [command options] <release bundle name> <release bundle version>"}
 
 const Arguments string = `	release bundle name
 		Release bundle name.
