+++ docs/artifactory/releasebundlesign/help.go
@@ -2,7 +2,7 @@ package releasebundlesign
 
 const Description = "Sign a release bundle."
 
-var Usage = []string{"jfrog rt rbs [command options] <release bundle name> <release bundle version>"}
+var Usage = []string{"jfrog ds rbs [command options] <release bundle name> <release bundle version>"}
 
 const Arguments string = `	release bundle name
 		Release bundle name.
