+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/jfrog/jfrog-cli/distribution"
 	"os"
 
 	corecommon "github.com/jfrog/jfrog-cli-core/docs/common"
@@ -127,6 +128,11 @@ func getCommands() []cli.Command {
 			Name:        cliutils.CmdXray,
 			Description: "Xray commands",
 			Subcommands: xray.GetCommands(),
+		},
+		{
+			Name:        cliutils.CmdDistribution,
+			Description: "Distribution commands",
+			Subcommands: distribution.GetCommands(),
 		},
 		{
 			Name:        cliutils.CmdCompletion,
